requirements.txt
streamlit
supabase
openai
google-genai
twilio
python-dotenv
pydantic